refactor: use psr-4
BREAKING CHANGE: switching to PSR-4 may impact the way to load the library
