+++ examples/example-create-02.php
@@ -15,7 +15,7 @@ ini_set( 'display_errors', 'on' );
 /**
  * Require CFPropertyList
  */
-require_once(__DIR__.'/../classes/CFPropertyList/CFPropertyList.php');
+require_once(__DIR__.'/../vendor/autoload.php');
 
 
 /*
@@ -43,7 +43,7 @@ $structure = array(
   'Kids Names' => array( 'John', 'Kyra' ),
 );
 
-$td = new CFTypeDetector();  
+$td = new CFTypeDetector();
 $guessedStructure = $td->toCFType( $structure );
 $plist->add( $guessedStructure );
