+++ src/CFPropertyList/CFBinaryPropertyList.php
