+++ tests/EmptyElementsTest.php
@@ -2,20 +2,11 @@
 
 namespace CFPropertyList;
 
-error_reporting(E_ALL | E_STRICT);
-ini_set('display_errors', 'on');
-
-if (!defined('LIBDIR')) {
-    define('LIBDIR', __DIR__ . '/../classes/CFPropertyList');
-}
-
-require_once(LIBDIR . '/CFPropertyList.php');
-
 class EmptyElementsTest extends \PHPUnit\Framework\TestCase {
     public function testWriteFile() {
         $expected = '<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple Computer//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0"><dict><key>string</key><string/><key>number</key><integer>0</integer><key>double</key><real>0</real></dict></plist>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0"><dict><key>string</key><string></string><key>number</key><integer>0</integer><key>double</key><real>0</real></dict></plist>
 ';
 
         $plist = new CFPropertyList();
